remove obsolete (int) cast
[[Imported from SVN: r1029]]

+++ istl/paamg/indicescoarsener.hh
@@ -338,7 +338,7 @@ namespace Dune
              aggregates[index->localIndexPair().local()] !=
              AggregatesMap<typename Graph::VertexDescriptor>::ISOLATED)
           {
-            assert(aggregates[index->localIndexPair().local()]<(int)attributes.size());
+            assert(aggregates[index->localIndexPair().local()]<attributes.size());
             assert(attributes[aggregates[index->localIndexPair().local()]] == std::numeric_limits<char>::max()
                    || attributes[aggregates[index->localIndexPair().local()]] == index->attribute());
             attributes[aggregates[index->localIndexPair().local()]] = index->attribute();
